mega-commit: implementation of basic instructions
This commit, made after this stage of the project was completed, further
expands on the simulate-able computer system. It also adds a
synthesizable `top.vhd` file, which forms the basis for FPGA
implementation, and relevant supporting VHDL entities (character
decoders and clock dividers) to support the computer system as it runs
within the FPGA fabric.